+++ src/app/components/home/home.component.ts
@@ -82,8 +82,7 @@ export class HomeComponent implements OnInit {
   }
 
   hideMenuOptions() {
-    this.g.wdLog(['hideMenuOptions']);
-    // this.g.isOpenMenuOptions = false;
+    this.g.wdLog(['hideMenuOptions A']);
     this.g.setParameter('isOpenMenuOptions', false, true);
   }