+++ src/app/providers/messaging.service.ts
@@ -683,14 +683,11 @@ export class MessagingService {
   /**
    *
    */
-  generateUidConversation(uid): string {
+  generateUidConversation(uid: string): string {
     this.firebaseMessagesKey = firebase.database().ref(this.urlMessages);
-    // creo il nodo conversazione generando un custom uid
     const newMessageRef = this.firebaseMessagesKey.push();
     const key = UID_SUPPORT_GROUP_MESSAGES + newMessageRef.key;
-    // sessionStorage.setItem(uid, key);
-    this.g.wdLog(['setItem ************** UID:', uid, ' KWY: ', key]);
-    // this.storageService.setItem(uid, key);
+    this.g.wdLog(['setItem ************** UID:', uid, ' KEY: ', key]);
     this.conversationWith = key;
     return key;
   }