+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # chat21-web-widget ver 3.0
 
+### 4.0.91
+- change: refactored 'send new message';
+- change: saved id conversation active in local storage;
+- change: open widget on last conversation;
+- bug fix: closed archived conversations;
+- bug-fix: startFromHome true/false management;
+- change: saved id active conversation on opening conversation detail;
+- bug-fix: changed workflow 'new conversation' button;
+
 ### 4.0.90
 - bug-fix: save new conversation in local storage